chore(git): Add '.gitignore' to prevent committing stuff that can be built
I made a big mistake in not adding this first. There are a lot of things that
Visual Code Studio generates that do not need to be added into the remote
repository, because anyone who clones the project can just build those generated
files automatically.

This lead to painful merge conflicts for cache files that are nearly impossible
to fix by hand, because those files should not have been committed in the first
place.

Signed-off-by: Jason Pena <jasonpena@awkless.com>