fix(vs): Use correct project ID, with hybrid build option
This solution file is basically a slimmed down version of the "SDK
solution.sln" file provided by the Prepar3D SDK samples folder when you
install the Prepar3D and SimConnect SDKs.

I assume Nate based his original solution file on the "SDK solution.sln"
file, but for whatever reason, it did not work properly. Apparently, we
need to have the hybrid build option setup, along with a "correct"
project ID. The project ID Nate used beforehand was not valid to Visual
Studio for some reason. So I just copied the ID for the "Managed Input
Request" project and it finally worked.

I have no idea why this worked, but I am rolling with it.

Signed-off-by: Alex <Alex@DESKTOP-LMCJ5MH>