ref(csproj): Migrate to package reference
Previous team used "packages.config" along with the "Reference" tag to
link package dependencies to the project. This method requires manual
package restoration, which can be a pain to do for new clones of the
project. Instead, I migrated the project to use "PackageReference",
which will not only perform automatic package restoration (if Visual
Studio is given permission to do so) and we no longer need to use the
"Reference" tag anymore, cleaning up and simplifying the code for the
build system configuration.

Signed-off-by: Alex <Alex@DESKTOP-LMCJ5MH>